pyproject.toml
[project]
name = "poc-diagrama-v2"
version = "1.0.0"
dependencies = [
  "fastapi==0.121.0",
  "uvicorn==0.38.0",
  "sqlmodel==0.0.27",
  "SQLAlchemy==2.0.44",
  "asyncpg==0.30.0",
  "greenlet==3.2.4",
  "python-json-logger==2.0.7",
  "psutil==6.1.0",
  "redis[hiredis]==5.2.1",  # ← Redis com hiredis (C parser, mais rápido)
  "pydantic-settings==2.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "pytest-cov>=7.0.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]